predicat checkpoint

+++ database/migrations/2023_05_30_140110_create_diagnose_table.php
@@ -17,10 +17,10 @@ class CreateDiagnoseTable extends Migration
             $table->id();
             $table->unsignedBigInteger('users_id');
             $table->foreign('users_id')->references('id')->on('users');
-            $table->integer('gds');
-            $table->integer('gdp');
-            $table->integer('tol_gluc');
-            $table->integer('HbA1C');
+            $table->float('gds');
+            $table->float('gdp');
+            $table->float('tol_gluc');
+            $table->float('HbA1C');
             $table->timestamps();
         });
     }
